+++ Cargo.toml
@@ -49,9 +49,8 @@ ndarray = "0.15.4"
 ndarray-rand = "0.14.0"
 ndarray-stats = " 0.5.0"
 rand_isaac = "0.3.0"
-plonky2 = { git = "https://github.com/recmo/plonky2", branch = "recmo/hacks" }
-
-#plonky2 = { git = "https://github.com/mir-protocol/plonky2", branch = "main" }
+#plonky2 = { git = "https://github.com/recmo/plonky2", branch = "recmo/hacks" }
+plonky2 = { git = "https://github.com/mir-protocol/plonky2", branch = "main" }
 
 [dev-dependencies]
 pretty_assertions = "1.0"
