+++ src/layers/maxpool.rs
@@ -1,6 +1,4 @@
 use ndarray::{s, Array3};
-// use ndarray::parallel::prelude::*;
-// required for finding element-wise maximum in array a
 use ndarray_stats::QuantileExt;
 
 pub struct MaxPool<T> {
@@ -47,5 +45,43 @@ pub fn max_pooling_layer(input: Array3<f32>, s: usize) -> MaxPool<f32> {
         n_params,
         n_multiplications,
         name: String::from("max-pool"),
+    }
+}
+
+#[cfg(test)]
+pub mod test {
+    use super::*;
+    use ndarray_rand::{rand::SeedableRng, rand_distr::Uniform, RandomExt};
+    use rand::rngs::StdRng;
+
+    #[test]
+    fn maxpool_test() {
+        let seed = 694201337;
+        let mut rng = StdRng::seed_from_u64(seed);
+
+        let input = Array3::random_using((116, 76, 32), Uniform::<f32>::new(-5.0, 5.0), &mut rng);
+        let s = 2;
+
+        let MaxPool::<f32> {
+            output: x,
+            n_params,
+            n_multiplications,
+            name,
+        } = max_pooling_layer(input, s);
+
+        assert_eq!(x.dim(), (58, 38, 32));
+
+        let (dim_x, dim_y, dim_z) = x.dim();
+
+        println!(
+            "
+        {} \n
+        # of parameters: {}\n
+        output dim: {}x{}x{}\n
+        # of ops: {}\n
+        output:\n
+        {}",
+            name, n_params, dim_x, dim_y, dim_z, n_multiplications, x
+        );
     }
 }
