+++ src/layers/mod.rs
@@ -1,0 +1,6 @@
+pub mod conv;
+pub mod flatten;
+pub mod fully_connected;
+pub mod maxpool;
+pub mod normalize;
+pub mod relu;
